+++ src/services/index.ts
@@ -4,6 +4,7 @@
 
 export { TableSelector } from './TableSelector';
 export { EffectsManager } from './EffectsManager';
+export { CritSuppression } from './CritSuppression';
 export { SaveManager } from './SaveManager';
 export { MidiQolHooks } from './MidiQolHooks';
 export { TableImporter } from './TableImporter';